+++ src/Web/Endpoint/Competitions/Competitions.php
@@ -5,6 +5,7 @@ namespace Wprs\Api\Web\Endpoint\Competitions;
 use Wprs\Api\Http\DownloaderInterface;
 use Wprs\Api\Web\Application;
 use Wprs\Api\Web\Endpoint\FilterInterface;
+use Wprs\Api\Web\System;
 
 /**
  * @phpstan-import-type apiData from \Wprs\Api\Web\Endpoint\ApiOutput
@@ -26,12 +27,38 @@ class Competitions extends Application
      */
     public function getData(?string $rankingDate = null): array
     {
-        $params = new CompetitionsParams();
-        $job = $this->getJob($rankingDate, $params);
+        $rankingDate = System::getRankingDate($rankingDate);
+        $results = $this->getBatch([$rankingDate]);
 
-        $results = parent::run([$job->getUrl()]);
-        $data = $results[0];
+        return $results[0];
+    }
 
-        return $job->getData($data);
+    /**
+     * @param array<string> $rankingDates
+     * @phpstan-return non-empty-array<apiData>
+     */
+    public function getBatch(array $rankingDates): array
+    {
+        $results = [];
+        $urls = [];
+        $jobs = [];
+
+        System::checkParams($rankingDates, System::PARAM_DATE);
+
+        foreach ($rankingDates as $rankingDate) {
+            $params = new CompetitionsParams();
+            $job = $this->getJob($rankingDate, $params);
+            $urls[] = $job->getUrl();
+            $jobs[] = $job;
+        }
+
+        $dataSets = parent::run($urls);
+
+        foreach ($dataSets as $index => $data) {
+            $job = $jobs[$index];
+            $results[] = $job->getData($data);
+        }
+
+        return $results;
     }
 }