+++ tests/Helpers/Builder/HtmlBuilder.php
@@ -15,13 +15,17 @@ use Wprs\Api\Web\Endpoint\Nations\Nations;
 use Wprs\Api\Web\Endpoint\Nations\NationsParams;
 use Wprs\Api\Web\Endpoint\Pilots\Pilots;
 use Wprs\Api\Web\Endpoint\Pilots\PilotsParams;
+use Wprs\Api\Web\Endpoint\Pilots\PilotsParser;
 use Wprs\Api\Web\System;
 
 class HtmlBuilder
 {
     private int $discipline;
     private int $regionId;
+    private ?int $nationId;
     private int $compId;
+    private int $pilotsCount;
+    private int $pilotsMax;
     private string $rankingDate;
     private Config $config;
     private HttpDownloader $downloader;
@@ -33,6 +37,7 @@ class HtmlBuilder
         $this->discipline = $this->config->getDiscipline();
         $this->rankingDate = $this->config->getRankingDate();
         $this->regionId = $this->config->getRegionId();
+        $this->nationId = $this->config->getNationId();
 
         $this->downloader = new HttpDownloader();
         $this->formatter = new HtmlFormatter();
@@ -48,7 +53,7 @@ class HtmlBuilder
             Utils::saveToFile($file, $html);
         }
 
-        $data = $this->config->getData($this->compId);
+        $data = $this->config->getData($this->pilotsCount, $this->pilotsMax, $this->compId);
         $configFile = Utils::getConfigFile();
 
         $json = json_encode($data, JSON_THROW_ON_ERROR | JSON_PRETTY_PRINT);
@@ -77,12 +82,24 @@ class HtmlBuilder
             $pages[$name] = HttpUtils::getResponseContent($response);
         }
 
+        list($this->pilotsCount, $this->pilotsMax) = $this->parsePilots($pages['pilots']);
         $this->compId = $this->parseCompetitions($pages['competitions']);
         $url = $this->getCompetitionUrl($this->compId);
         $response = $this->downloader->get($url);
         $pages['competition'] = HttpUtils::getResponseContent($response);
 
         return $pages;
+    }
+
+    /**
+     * @return array{0: int, 1: int}
+     */
+    private function parsePilots(string $html): array
+    {
+        $parser = new PilotsParser();
+        $data = $parser->parse($html);
+
+        return [$data->getOverallCount(), $data->getItemCount()];
     }
 
     private function parseCompetitions(string $html): int
@@ -137,7 +154,7 @@ class HtmlBuilder
     private function getPilotsUrl(): string
     {
         $path = System::getPath($this->discipline, Pilots::class);
-        $params = new PilotsParams($this->regionId);
+        $params = new PilotsParams($this->regionId, $this->nationId);
         $query = $params->getQueryParams($this->rankingDate);
 
         return HttpUtils::buildQuery($query, $path);